swal added

+++ src/pages/Toys/MyToysRow.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect } from 'react';
 import { Link, useActionData } from 'react-router-dom';
 import sortBy from 'sort-by';
+import Swal from 'sweetalert2';
 
 const MyToysRow = ({myToy, myToys, setMyToys}) => {
   console.log(myToy);
@@ -9,17 +10,34 @@ const MyToysRow = ({myToy, myToys, setMyToys}) => {
   
    const handleDelete = id =>{
 
-      fetch(`https://toy-center-server.vercel.app/toys/${id}`,{
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          fetch(`https://toy-center-server.vercel.app/toys/${id}`,{
         method:"DELETE",
       })
       .then(res => res.json())
       .then(data =>{
-        console.log(data);
         if(data.deletedCount > 0){
+          Swal.fire(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success'
+          )
           const reaming = myToys.filter(myToy => myToy._id !== id)
           setMyToys(reaming)
         }
       })
+        }
+      })
+
    }
     return (
       <tr>
